+++ NEWS.md
@@ -1,3 +1,5 @@
+# bigrquery (development version)
+
 # bigrquery 1.3.1
 
 * Now requires gargle 0.5.0