+++ NEWS.md
@@ -1,5 +1,10 @@
 # bigrquery (development version)
 
+* `bq_table_download()` works once again with large row counts
+  (@gjuggler, #395). Google's API has stopped accepting `startIndex`
+  parameters with scientific formatting, which was happening for large
+  values (>1e5) by default.
+
 # bigrquery 1.3.1
 
 * Now requires gargle 0.5.0